Split up src/sources/util/unix.rs to satisfy component checks
