+++ src/sources/util/mod.rs
@@ -3,8 +3,10 @@ mod http;
 pub mod multiline_config;
 #[cfg(all(feature = "sources-utils-tls", feature = "listenfd"))]
 mod tcp;
-#[cfg(all(unix, feature = "sources-utils-unix",))]
-mod unix;
+#[cfg(all(unix, feature = "sources-socket"))]
+mod unix_datagram;
+#[cfg(all(unix, feature = "sources-utils-unix"))]
+mod unix_stream;
 
 #[cfg(any(feature = "sources-http", feature = "sources-logplex"))]
 pub(crate) use self::http::add_query_parameters;
@@ -13,5 +15,7 @@ pub(crate) use self::http::{ErrorMessage, HttpSource, HttpSourceAuthConfig};
 pub use multiline_config::MultilineConfig;
 #[cfg(all(feature = "sources-utils-tls", feature = "listenfd"))]
 pub use tcp::{SocketListenAddr, TcpSource};
+#[cfg(all(unix, feature = "sources-socket",))]
+pub use unix_datagram::build_unix_datagram_source;
 #[cfg(all(unix, feature = "sources-utils-unix",))]
-pub use unix::{build_unix_datagram_source, build_unix_stream_source};
+pub use unix_stream::build_unix_stream_source;
